--- a/util/httputil/compression.go
+++ b/util/httputil/compression.go
@@ -31,18 +31,18 @@ const (
 
 // Wrapper around http.Handler which adds suitable response compression based
 // on the client's Accept-Encoding headers.
-type compressedResponseWriter struct {
+type deflatedResponseWriter struct {
 	http.ResponseWriter
 	writer io.Writer
 }
 
 // Writes HTTP response content data.
-func (c *compressedResponseWriter) Write(p []byte) (int, error) {
+func (c *deflatedResponseWriter) Write(p []byte) (int, error) {
 	return c.writer.Write(p)
 }
 
-// Closes the compressedResponseWriter and ensures to flush all data before.
-func (c *compressedResponseWriter) Close() {
+// Closes the deflatedResponseWriter and ensures to flush all data before.
+func (c *deflatedResponseWriter) Close() {
 	if zlibWriter, ok := c.writer.(*zlib.Writer); ok {
 		zlibWriter.Flush()
 	}
@@ -51,10 +51,10 @@ func (c *compressedResponseWriter) Close() {
 	}
 }
 
-// Constructs a new compressedResponseWriter based on client request headers.
-func newCompressedResponseWriter(writer http.ResponseWriter) *compressedResponseWriter {
+// Constructs a new deflatedResponseWriter based on client request headers.
+func newDeflateResponseWriter(writer http.ResponseWriter) *deflatedResponseWriter {
 	writer.Header().Set(contentEncodingHeader, deflateEncoding)
-	return &compressedResponseWriter{
+	return &deflatedResponseWriter{
 		ResponseWriter: writer,
 		writer:         zlib.NewWriter(writer),
 	}
@@ -75,7 +75,7 @@ func (c CompressionHandler) ServeHTTP(writer http.ResponseWriter, req *http.Requ
 			gzhttp.GzipHandler(c.Handler).ServeHTTP(writer, req)
 			return
 		case deflateEncoding:
-			compWriter := newCompressedResponseWriter(writer)
+			compWriter := newDeflateResponseWriter(writer)
 			c.Handler.ServeHTTP(compWriter, req)
 			compWriter.Close()
 			return