@roidelapluie 2023-05-16
See also https://github.com/prometheus/prometheus/pull/10782