@alanprot 2023-05-16
Nice! A similar PR, but here it's using the `GzipHandler` directly which has other optimizations:

See: https://github.com/klauspost/compress/blob/9a951bee0e5e9df261e6ca6a68bfa1e819f333d0/gzhttp/compress.go#L420-L432